requirements.txt
# Core dependencies for Persona Drift Detection Pipeline
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
datasets>=2.14.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.10.0
pyarrow>=12.0.0
fastparquet>=2023.4.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Topic modeling
gensim>=4.3.0
bertopic>=0.15.0
umap-learn>=0.5.3
hdbscan>=0.8.29

# NLP
nltk>=3.8
spacy>=3.6.0
wordcloud>=1.9.0
